feat: AI 校正分作为最终分 + 两位小数 + 炫耀图
1) AI 人工修正(±10)现在真正叠加到最终分:
   - roast 流首行输出 `META {json}`(AI 校正后的 final_score/tier/delta/percentile),
     前端据此把大字分数从脚本初评更新为 AI 复核后的值,并显示「脚本初评 X · AI +N」
   - 协议: 模型先输出一行 `@@ADJUST <delta>@@`,服务端解析(clamp ±10)、算 adjusted、
     落库、算百分位,再流式输出报告(已剥离该控制行)
   - 反作弊: 仅默认模型的校正分写入公开榜,BYO Key 不写榜(防自抬排名)
2) 分数改为两位小数(score.ts: final 不再取整;tierFor/clampScore 抽出复用;
   DB final_score 改 REAL;榜单/卡片 toFixed(2)),便于精排
3) 新增「📸 保存炫耀图」: ShareCard 离屏卡片 + html-to-image 导出 PNG(头像转 dataURL 避免污染画布)
4) 百分位/落库从 scan 路由移到 roast 路由(那里才有 AI 最终分),scan 回归纯确定性结果

实测(真实模型): gaearon 98.00/夯, META 首行正确解析, 落榜正确, delta 解析含 +N 边界。
17 测试全过(fixtures 改 2 位小数), typecheck/lint/build 全绿。

+++ src/lib/db.ts
@@ -36,7 +36,7 @@ function ensureSchema(db: Client): Promise<void> {
              display_name TEXT,
              avatar_url   TEXT,
              profile_url  TEXT,
-             final_score  INTEGER NOT NULL,
+             final_score  REAL NOT NULL,
              tier         TEXT NOT NULL,
              hidden       INTEGER NOT NULL DEFAULT 0,
              scanned_at   INTEGER NOT NULL
